fix(AlertGateway.kt) : 파일 정리
