+++ app/src/main/java/com/iron/espresso/presentation/home/alert/AlertListActivity.kt
@@ -6,89 +6,13 @@ import android.graphics.Rect
 import android.os.Bundle
 import android.view.MenuItem
 import android.view.View
-import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.recyclerview.widget.RecyclerView
 import com.iron.espresso.R
 import com.iron.espresso.base.BaseActivity
 import com.iron.espresso.databinding.ActivityAlertListBinding
-import com.iron.espresso.domain.entity.AlertType
 import com.iron.espresso.ext.dp
-import com.iron.espresso.presentation.home.apply.ApplyStudyActivity
-import com.iron.espresso.presentation.home.mystudy.MyStudyDetailActivity
 import dagger.hilt.android.AndroidEntryPoint
-
-object AlertGateway {
-
-    fun goToPage(context: Context, alertType: AlertType, studyId: Int, studyTitle: String = "") {
-        when (alertType) {
-            AlertType.APPLY_NEW -> {
-                context.startActivity(ApplyStudyActivity.getIntent(context, studyId = studyId))
-            }
-            AlertType.STUDY_UPDATE -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.STUDY_DELEGATE -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.STUDY_DELETE -> {
-                Toast.makeText(context, "$studyTitle 스터디가 삭제되었습니다.", Toast.LENGTH_SHORT).show()
-            }
-            AlertType.APPLY_ALLOW -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.APPLY_REJECT -> {
-                Toast.makeText(context, "$studyTitle 스터디 입장이 거절되었습니다.", Toast.LENGTH_SHORT).show()
-            }
-            AlertType.NOTICE_NEW -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.NOTICE_UPDATE -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.CHAT -> {
-            }
-            AlertType.PUSH_TEST -> {
-            }
-        }
-    }
-
-
-    fun goToPageByPush(context: Context, alertType: AlertType, studyId: Int, alertId: Int) {
-        when (alertType) {
-            AlertType.APPLY_NEW -> {
-                context.startActivity(ApplyStudyActivity.getIntent(context, studyId = studyId))
-            }
-            AlertType.STUDY_UPDATE -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.STUDY_DELEGATE -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.STUDY_DELETE -> {
-                context.startActivity(AlertListActivity.getIntent(context))
-            }
-            AlertType.APPLY_ALLOW -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.APPLY_REJECT -> {
-                context.startActivity(AlertListActivity.getIntent(context))
-            }
-            AlertType.NOTICE_NEW -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.NOTICE_UPDATE -> {
-                context.startActivity(MyStudyDetailActivity.getInstance(context, id = studyId))
-            }
-            AlertType.CHAT -> {
-            }
-            AlertType.PUSH_TEST -> {
-                context.startActivity(AlertListActivity.getIntent(context))
-            }
-        }
-    }
-}
-
 
 @AndroidEntryPoint
 class AlertListActivity : BaseActivity<ActivityAlertListBinding>(R.layout.activity_alert_list) {
