+++ app/src/main/res/values/strings.xml
@@ -165,4 +165,6 @@
 
     <string name="empty_delegate_member">위임할 멤버가 존재하지 않아요</string>
     <string name="check">확인하기</string>
+    <string name="toast_study_deleted">%s 스터디가 삭제되었습니다.</string>
+    <string name="toast_study_rejected">%s 스터디 입장이 거절되었습니다.</string>
 </resources>
